Implementando funções no site

+++ Speckoz.UniLinks/UniLinks.API/Controllers/DisciplinesController.cs
@@ -19,11 +19,13 @@ namespace UniLinks.API.Controllers
 	{
 		private readonly IDisciplineBusiness _disciplineBusiness;
 		private readonly ICourseBusiness _courseBusiness;
+		private readonly IStudentBusiness _studentBusiness;
 
-		public DisciplinesController(IDisciplineBusiness disciplineBusiness, ICourseBusiness courseBusiness)
+		public DisciplinesController(IDisciplineBusiness disciplineBusiness, ICourseBusiness courseBusiness, IStudentBusiness studentBusiness)
 		{
 			_disciplineBusiness = disciplineBusiness;
 			_courseBusiness = courseBusiness;
+			_studentBusiness = studentBusiness;
 		}
 
 		[HttpPost("add")]
@@ -72,6 +74,21 @@ namespace UniLinks.API.Controllers
 					return Ok(discs);
 
 				return NotFound("Uma ou todas as disciplinas requisitadas nao foram localizadas!");
+			}
+
+			return BadRequest();
+		}
+
+		[HttpGet("{disciplineId}")]
+		[Authorizes]
+		public async Task<IActionResult> GetDisciplineByDisciplineId([Required] Guid disciplineId)
+		{
+			if (ModelState.IsValid)
+			{
+				if (await _disciplineBusiness.FindByDisciplineIdTaskAsync(disciplineId) is DisciplineVO discipline)
+					return Ok(discipline);
+
+				return NotFound("Nao existe nenhuma disciplina com o Id informado");
 			}
 
 			return BadRequest();
@@ -130,7 +147,7 @@ namespace UniLinks.API.Controllers
 					return BadRequest("É necessario informar a sala!");
 
 				if (await _disciplineBusiness.UpdateTaskAync(newDiscipline) is DisciplineVO disciplineUpdated)
-					return Ok(disciplineUpdated);
+					return Created($"/Disciplines/{disciplineUpdated.DisciplineId}", disciplineUpdated);
 			}
 
 			return BadRequest();
@@ -138,17 +155,24 @@ namespace UniLinks.API.Controllers
 
 		[HttpDelete("{disciplineId}")]
 		[Authorizes(UserTypeEnum.Coordinator)]
-		public async Task<IActionResult> DeleteTaskAsync([Required] Guid disciplineId)
+		public async Task<IActionResult> DeleteTaskAsync([Required] Guid disciplineId, [FromServices] ILessonBusiness lessonBusiness)
 		{
 			if (ModelState.IsValid)
 			{
 				if (!(await _disciplineBusiness.FindByDisciplineIdTaskAsync(disciplineId) is DisciplineVO discipline))
 					return NotFound("Nao existe uma disciplina com esse Id");
 
-				CourseVO course = await _courseBusiness.FindByCourseIdTaskAsync(discipline.CourseId);
+				if (!(await _courseBusiness.FindByCourseIdTaskAsync(discipline.CourseId) is CourseVO course))
+					return NotFound("Nao existe nenhum curso com o coordenador informado");
 
 				if (course.CoordinatorId != Guid.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
 					return Unauthorized("Voce nao tem autorizaçao para remover uma disciplina de outro curso!");
+
+				if (await lessonBusiness.ExistsByDisciplineIdTaskAsync(disciplineId))
+					return BadRequest("Nao é possivel excluir a disciplina, pois existem aulas utilizando-a!");
+
+				if (await _studentBusiness.ExistsStudentWithDisciplineTaskAsync(disciplineId))
+					return BadRequest("Nao é possivel excluir a disciplina, pois existem alunos utilizando-a!");
 
 				await _disciplineBusiness.DeleteTaskAsync(discipline.DisciplineId);
 				return NoContent();
